requirements.txt
PyQt6
PyMuPDF
pytesseract
sqlalchemy
anthropic
pillow
